fix deprecations
- ioutil.TempFIle deprecated
- golang 1.15 deprecated
